refactor folder structure.
move DataService and InventoryService to app folder for go install to reference
move DataServiceAdmin and InventoryServiceAdmin to tool folder for go install to reference
move libraries of DataService to src/DataService folder as internal none reference-able code
move libraries of InventoryService to src/Inventory folder as internal none reference-able code
