+++ src/InventoryService/go.mod
@@ -22,6 +22,6 @@
 // ******************************************************************************************************************
 
 
-module UniTao/InventoryService
+module InventoryService
 
 go 1.18
