+++ test/data/DataService01/initTable.sh
@@ -28,13 +28,13 @@ SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]:-$0}"; )" &> /dev/null && 
 pushd $SCRIPT_DIR/../../../
 
 # reset table structure
-go run ./src/DataServiceAdmin/main.go table \
+go run ./tool/DataServiceAdmin/main.go table \
     -config $SCRIPT_DIR/config.json \
     -table ./dbSchemas/DynamoDB/Tables.json \
     -reset true
 
 # import schema record
-go run ./src/DataServiceAdmin/main.go data \
+go run ./tool/DataServiceAdmin/main.go data \
     -config $SCRIPT_DIR/config.json \
     -data ./lib/Schema/data/schema.json
 popd
