+++ test/data/DataService02/importData.sh
@@ -27,7 +27,7 @@ SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]:-$0}"; )" &> /dev/null && 
 
 pushd $SCRIPT_DIR/../../../
 
-go run ./src/DataServiceAdmin/main.go data \
+go run ./tool/DataServiceAdmin/main.go data \
     -config $SCRIPT_DIR/config.json \
     -data ./test/data/SampleData/SeparateData/DS_02.json
